--- vuewebapp/app/templates/predict.html
@@ -1,89 +1,0 @@
-{% extends "base.html" %}
-
-{% block main %}
-<!-- Modal -->
-<div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
-    <div class="modal-dialog" role="document">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="exampleModalLabel">Prediction Results</h5>
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                    <span aria-hidden="true">&times;</span>
-                </button>
-            </div>
-            <div class="modal-body">
-                <h5>Prediction Input:</h5>
-                <div>Sepal Length: <span id="sl"></span></div>
-                <div>Sepal Width: <span id="sw"></span></div>
-                <div>Petal Length: <span id="pl"></span></div>
-                <div>Petal width: <span id="pw"></span></div>
-                <h5 class="pt-3">Prediction Classification:</h5>
-                <div id="prediction"></div>
-            </div>
-            <div class="modal-footer">
-                <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                <a class="btn btn-primary" href="/results" role="button">View DB</a>
-            </div>
-        </div>
-    </div>
-</div>
-
-<div class="container pt-5">
-    <div class="row justify-content-md-center">
-        <div class="col-md-4">
-            <h1>Iris Prediction</h1>
-            <form action="" id="post-form">
-                {% csrf_token %}
-                <div class="form-group">
-                    <label for="sepal_length">Sepal Length</label>
-                    <input type="number" step="0.1" class="form-control" id="sepal_length" placeholder="" required>
-                </div>
-                <div class="form-group">
-                    <label for="Sepal Width">Sepal Width</label>
-                    <input type="number" step="0.1" class="form-control" id="sepal_width" placeholder="" required>
-                </div>
-                <div class="form-group">
-                    <label for="petal_length">Petal Length</label>
-                    <input type="number" step="0.1" class="form-control" id="petal_length" placeholder="" required>
-                </div>
-                <div class="form-group">
-                    <label for="petal_width">Petal Width</label>
-                    <input type="number" step="0.1" class="form-control" id="petal_width" placeholder="" required>
-                </div>
-                <button type="submit" value="Submit" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal">Submit</button>
-            </form>
-        </div>
-    </div>
-</div>
-
-<script>
-    $(document).on('submit', '#post-form',function(e){
-        e.preventDefault();
-        $.ajax({
-            type:'POST',
-            url:'{% url "predict:submit_prediction" %}',
-            data:{
-                sepal_length:$('#sepal_length').val(),
-                sepal_width:$('#sepal_width').val(),
-                petal_length:$('#petal_length').val(),
-                petal_width:$('#petal_width').val(),
-                csrfmiddlewaretoken:$('input[name=csrfmiddlewaretoken]').val(),
-                action: 'post'
-            },
-            success:function(json) {
-                document.forms["post-form"].reset();
-                document.getElementById("prediction").innerHTML = json['result']
-                document.getElementById("sl").innerHTML = json['sepal_length']
-                document.getElementById("sw").innerHTML = json['sepal_width']
-                document.getElementById("pl").innerHTML = json['petal_length']
-                document.getElementById("pw").innerHTML = json['petal_width']
-            },
-            error : function(xhr,errmsg,err) {
-
-            }
-        });
-    })
-</script>
-
-{% endblock %}
-
