--- vuewebapp/app/templates/results.html
@@ -1,36 +1,0 @@
-{% extends "base.html" %}
-
-{% block main %}
-    <div class="container pt-5">
-        <div class="row">
-
-        <h1>Prediction Results</h1>
-
-            <table class="table">
-                <thead>
-                <tr>
-                <th scope="col">#</th>
-                <th scope="col">Sepal length</th>
-                <th scope="col">Sepal width</th>
-                <th scope="col">Petal length</th>
-                <th scope="col">Petal width</th>
-                <th scope="col">Prediction</th>
-                </tr>
-                </thead>
-                <tbody>
-                {% for data in dataset %}
-                    <tr>
-                    <th scope="row">{{ data.id }}</th>
-                    <td>{{ data.sepal_length }}</td>
-                    <td>{{ data.sepal_width }}</td>
-                    <td>{{ data.petal_length }}</td>
-                    <td>{{ data.petal_width }}</td>
-                    <td>{{ data.classification }}</td>
-                    </tr>
-                {% endfor %}
-
-                </tbody>
-            </table>
-        </div>
-    </div>
-{% endblock %}
